+++ src/test/kotlin/de/bdr/servko/keycloak/gematik/idp/model/GematikIDPConfigTest.kt
@@ -1,12 +1,13 @@
 /*
- *  Copyright 2023 Bundesdruckerei GmbH and/or its affiliates
- *  and other contributors.
+ * Copyright 2025 Bundesdruckerei GmbH and/or its affiliates
+ * and other contributors.
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
  *  You may obtain a copy of the License at
  *
  *  http://www.apache.org/licenses/LICENSE-2.0
+ *
  *
  *  Unless required by applicable law or agreed to in writing, software
  *  distributed under the License is distributed on an "AS IS" BASIS,
@@ -209,6 +210,20 @@ internal class GematikIDPConfigTest {
 
         // assert
         assertThat(result).isEqualTo(multipleIdentityMode)
+    }
+
+    @ParameterizedTest
+    @ValueSource(booleans = [true, false])
+    fun setAuthenticatorAutoLaunch(authenticatorAutoLaunch: Boolean) {
+        // arrange
+        val idpConfig = initGematikIDPConfig()
+        idpConfig.setAuthenticatorAutoLaunch(authenticatorAutoLaunch)
+
+        // act
+        val result = idpConfig.getAuthenticatorAutoLaunch()
+
+        // assert
+        assertThat(result).isEqualTo(authenticatorAutoLaunch)
     }
 
     @ParameterizedTest